reorganise project files and add router
