+++ src/App.tsx
@@ -1,9 +1,21 @@
 import React from "react";
-import AddLevel from "./components/AddLevel"
+
+import AddLevelModal from "./components/AddLevelModal";
+import CommissionTypeModal from "./components/CommissionTypeModal";
 import "./App.css";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 function App() {
-  return <AddLevel />;
+  return (
+    <Router>
+      <Switch>
+        <Route path={"/add-level"} component={AddLevelModal} />
+        <Route path={"/commission-type"} >
+          <CommissionTypeModal levelName="Senior Agent" />
+          </Route>
+      </Switch>
+    </Router>
+  );
 }
 
 export default App;
